requirements.txt
requests~=2.22.0
django~=2.2.9
pytest~=5.4.3
pytest-django
djangorestframework~=3.11.1
djangorestframework-simplejwt

django-filter~=2.4.0
py~=1.8.1
PyJWT~=1.7.1
cryptography~=2.9.2
pip~=19.3.1
wheel~=0.33.6
attrs~=19.3.0
pytz~=2019.3
toml~=0.10.1
sqlparse~=0.3.0
Jinja2~=2.11.2
wcwidth~=0.1.8
pluggy~=0.13.1
certifi~=2019.9.11
chardet~=3.0.4
urllib3~=1.25.6
idna~=2.8
pyparsing~=2.4.6
six~=1.14.0
setuptools~=42.0.2
packaging~=20.1
zipp~=2.2.0